% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagen.gp.uni}
\alias{datagen.gp.uni}
\title{Datagen for GP.}
\usage{
datagen.gp.uni(N, lambda_loc, lambda_sca_std, nu_loc, nu_sca, fun)
}
\arguments{
\item{N}{Number of persons.}

\item{lambda_loc}{(Standardized) Loadings for theta_loc.}

\item{lambda_sca_std}{Standardized loadings for theta_sca.}

\item{nu_loc}{Intercepts for location models.}

\item{nu_sca}{Intercepts for scale models.}

\item{fun}{True function for theta_sca = fun(theta_loc)}
}
\value{
List of params, meta-data (meta), and stan data (data).
}
\description{
Generates data for univariate GP model.
}
\details{
Generates dataset for GP model.
The (univariate) GP model is such that theta_sca = fun(theta_loc).
}
\author{
Stephen R. Martin
}
\keyword{internal}
