% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.cond_mvn}
\alias{.cond_mvn}
\title{Compute conditional MVN.}
\usage{
.cond_mvn(x, mu, sigma)
}
\arguments{
\item{x}{Column vectors (Matrix) of conditional values (conditional variables).}

\item{mu}{Vector of means (all variables).}

\item{sigma}{Covariance matrix (all variables).}
}
\description{
Computes conditional MVN.
}
\details{
All (Nx(Q-p)) values are conditioned on the (Nxp) values in x.
That is, if sigma is 10x10, and mu is 10, and x is Nx4, then the last Nx6 are conditioned on x.
}
\author{
Stephen Martin
}
