% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.extract_transform}
\alias{.extract_transform}
\title{Extract and transform samples to R vectors, matrices, arrays.}
\usage{
.extract_transform(
  object,
  par,
  nsamples = NULL,
  transpose = FALSE,
  stan = FALSE,
  ...
)
}
\arguments{
\item{object}{stanfit.}

\item{par}{Character. Name of parameter to extract and transform.}

\item{nsamples}{Numeric (Optional). Number of samples to use.}

\item{transpose}{Logical (Default: FALSE). Only applicable if two dimensional.
Intended for use when an array of vectors is used.
When TRUE, will transpose such that each row is a vector element, and each column is a vector.}

\item{stan}{Logical (Default: FALSE). If true, will not repermute; will be in same order as stan provides.}

\item{...}{Not used.}
}
\value{
If par is a vector, then a PxS array.
If par is matrix, then a RxCxS array.
If par is array of matrices, then RxCxA1xA2x...xS array.
If par is array of vectors, then a AxRxS array (unless transpose = TRUE).
}
\description{
Extracts samples from stan fit, converts to R-friendly vectors, matrices, arrays, for each sample.
}
\author{
Stephen R. Martin
}
\keyword{internal}
