% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{.get_latent_vars}
\alias{.get_latent_vars}
\title{Compute latent variances from fitted.}
\usage{
.get_latent_vars(object, prob, SD = FALSE, summary = TRUE)
}
\arguments{
\item{object}{omegad object.}

\item{prob}{Probability mass for interval.}

\item{SD}{Logical (Default: FALSE). Whether to compute SDs (TRUE) or Vars (FALSE).}

\item{summary}{Logical (Default: TRUE). Whether to summarize (TRUE) or return samples (FALSE).}
}
\value{
Named vector of latent variances.
}
\description{
Computes latent variances from fitted.
}
\details{
Standard decomp rules could be followed.
GPs add alpha^2 marginal variance, and the linear functions follow standard path tracing rules.
However, a more straightforward approach is to just directly compute the variance from the fitted scores, which uses fewer assumptions and inherits all uncertainty from the params anyway.
}
\author{
Stephen R. Martin
}
\keyword{internal}
