% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{.get_loadings_matrix}
\alias{.get_loadings_matrix}
\title{Create FxJ loadings matrix, padded with zeroes}
\usage{
.get_loadings_matrix(formList, mm)
}
\arguments{
\item{formList}{Formula list (one entry per factor)}

\item{mm}{Model matrix}

\item{names}{Output of .get_names_formulaList}
}
\value{
FxJ matrix with the indicator indices for each factor. Padded with zeroes.
}
\description{
Takes (RHS) formula list and model *matrix*
}
\keyword{internal}
