% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{.get_names_formulaList}
\alias{.get_names_formulaList}
\title{Takes formula and returns names}
\usage{
.get_names_formulaList(formList, formula = FALSE)
}
\arguments{
\item{formList}{List of 2-sided Formulas}

\item{formula}{Logical (Default: FALSE). If TRUE, returns the rhs formula terms (e.g., I(x^2), rather than x)}
}
\description{
Takes formula and returns names
}
\keyword{internal}
