% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{.lavFit}
\alias{.lavFit}
\title{Fit similar model in lavaan.}
\usage{
.lavFit(object, ...)
}
\arguments{
\item{object}{omegad object.}
}
\value{
lavaan fit
}
\description{
Fit similar model in lavaan.
}
\author{
Stephen R. Martin
}
\keyword{internal}
