% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{.lavOmega}
\alias{.lavOmega}
\title{Compute omegas using lavaan.}
\usage{
.lavOmega(object)
}
\arguments{
\item{object}{omegad object.}
}
\value{
Reliability output from \code{\link[semTools]{reliability}}.
}
\description{
Compute omegas using lavaan.
}
\author{
Stephen R. Martin
}
\keyword{internal}
