% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.parse_formula.predict}
\alias{.parse_formula.predict}
\title{Create data structures needed for prediction.}
\usage{
.parse_formula.predict(object, newdata)
}
\arguments{
\item{object}{omegad object.}

\item{newdata}{Data.frame. Must contain the exogenous (if used) and latent factor names.}
}
\value{
List of data structures for prediction.
}
\description{
Converts object and newdata into appropriate design matrices.
}
\author{
Stephen R. Martin
}
\keyword{internal}
