% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{.parse_formula}
\alias{.parse_formula}
\title{Parses formula and data into stan_data structure}
\usage{
.parse_formula(formula, data, rescale = TRUE, ...)
}
\arguments{
\item{formula}{Formula or list of formulas with factor name on LHS, indicators on RHS.}

\item{data}{data.frame containing indicators.}

\item{rescale}{Logical. Whether to standardize indicators.}

\item{...}{Currently not used.}
}
\value{
stan_data list.
}
\description{
Parses formula and data into stan_data structure
}
\keyword{internal}
