% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{.standardize_output}
\alias{.standardize_output}
\title{Standardize summary output.}
\usage{
.standardize_output(x)
}
\arguments{
\item{x}{summary.omegad object.}
}
\value{
Summary object (with standardized values).
}
\description{
Recompute parameter outputs as standardized values.
}
\details{
Loadings and intercepts can be recomputed assuming standardized values via linear transformations.
For loadings:
\eqn{\beta_{a}SD_a = \beta_{b}SD_b}
\eqn{\beta_{un}SD_{un} = \beta_{st}}
\eqn{\beta_{0c} = \beta_{0} + \beta_1\bar x}
\eqn{\Lambda^c = D\Lambda}
If a standardized endogenous variable (Var(theta) = 1), then exogenous predictors have:
\eqn{\beta_{st} = \beta_{un}/SD_{latent}}
\eqn{\sigma^2_\epsilon = 1/Var(latent)}

For accuracy, I think we need to compute these over each posterior sample.
}
\author{
Stephen R. Martin
}
\keyword{internal}
