% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.omegad}
\alias{fitted.omegad}
\title{Extract fitted error factor scores, omega1, and omega2 values.}
\usage{
\method{fitted}{omegad}(object, summary = TRUE, prob = 0.95, ...)
}
\arguments{
\item{object}{omegad object.}

\item{summary}{Logical (Default: TRUE). Whether to return summary or array of samples.}

\item{prob}{Numeric (Default: .95). The amount of probability mass to include within the credible. Default values provide a 95\% credible interval.}

\item{...}{Unused.}
}
\value{
List of arrays. The list contains arrays of posterior summaries (\code{summary = TRUE}) or samples (\code{summary = FALSE}) for the predicted error factor, omega1, and omega2.
If \code{summary = TRUE}, then the arrays are of size [N, 4, F], where N is the number of predicted observations, the four columns correspond to the summary statistics (mean, sd, and intervals), and F is the number of factors.
The factor dimension is named, allowing for \code{out$omega1[,,"agreeableness"]}.
If \code{summary = FALSE}, then the arrays are of size [N, F, S], where S is the number of samples (\code{nsamples}).
}
\description{
Extracts estimated Error factor scores and reliability estimates for the fitted data.
}
\author{
Stephen R. Martin
}
