% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{\%IfNull\%}
\alias{\%IfNull\%}
\title{Operator for testing NULL and returning expr if NULL}
\usage{
object \%IfNull\% expr
}
\arguments{
\item{object}{Object to test NULL on.}

\item{expr}{Expression to evaluate.}
}
\value{
object if not NULL, results of expression if object is NULL
}
\description{
Operator for testing NULL and returning expr if NULL
}
\author{
Stephen R Martin
}
\keyword{internal}
