% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.omegad}
\alias{loo.omegad}
\title{omegad method for computing leave-one-out CV scores.}
\usage{
\method{loo}{omegad}(object, ...)
}
\arguments{
\item{object}{omegad object.}

\item{...}{}
}
\value{
loo object.
}
\description{
LOO method for omegad objects.
}
\details{
LOO method for omegad objects.
}
\author{
Stephen R. Martin
}
