% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nlist}
\alias{nlist}
\title{Construct named list from arguments.}
\usage{
nlist(...)
}
\arguments{
\item{...}{Objects for list.}
}
\value{
Named list.
}
\description{
Construct named list from arguments.
}
\author{
Stephen R Martin
}
\keyword{internal}
