% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstantools.R
\name{nsamples.omegad}
\alias{nsamples.omegad}
\title{Extract number of posterior samples.}
\usage{
nsamples.omegad(object)
}
\arguments{
\item{object}{omegad object.}
}
\value{
Integer. How many posterior post-warmup samples were drawn.
}
\description{
Extracts the number of posterior samples tored in a fitted omegad model.
}
\author{
Stephen R. Martin
}
\keyword{internal}
