% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{omega}
\alias{omega}
\title{Fit Omega model}
\usage{
omega(formula, data, ...)
}
\arguments{
\item{formula}{Formula specifying the factor structure. See details.}

\item{data}{Data frame containing indicators as columns.}

\item{...}{Options passed onto sampling.}
}
\value{
omegad object.
}
\description{
Fit Omega model
}
