% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omegad-package.R
\docType{package}
\name{omegad-package}
\alias{omegad-package}
\title{The 'omegad' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephensrmmartin/omegad}
  \item Report bugs at \url{https://github.com/stephensrmmartin/omegad/issues}
}

}
\author{
\strong{Maintainer}: Stephen R. Martin \email{stephenSRMMartin@gmail.com}

Authors:
\itemize{
  \item Donald R. Williams \email{drwwilliams@ucdavis.edu}
  \item Philippe Rast \email{prast@ucdavis.edu}
}

}
