% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.omegad}
\alias{predict.omegad}
\title{Predict method for omegad object.}
\usage{
\method{predict}{omegad}(object, newdata, summary = TRUE, prob = 0.95,
  nsamples = NULL, error = TRUE, ...)
}
\arguments{
\item{object}{omegad object.}

\item{newdata}{Data.frame. Must contain the exogenous (if used) and latent factor names.}

\item{summary}{Logical (Default: TRUE). Whether to return summary or array of samples.}

\item{prob}{Numeric (Default: .95). The amount of probability mass to include within the credible interval. Default values provide a 95\% credible interval.}

\item{nsamples}{Numeric (Default: All). Number of posterior samples to use.}

\item{error}{Logical (Default: TRUE). Whether the predicted latent variables are sampled with stochastic error or not.}

\item{...}{Unused.}

\item{samples}{Numeric (Default: All). Number of posterior samples to use.}
}
\value{
List of arrays. The list contains arrays of posterior summaries (\code{summary = TRUE}) or samples (\code{summary = FALSE}) for the predicted error factor, omega1, and omega2.
If \code{summary = TRUE}, then the arrays are of size [N, 4, F], where N is the number of predicted observations, the four columns correspond to the summary statistics (mean, sd, and intervals), and F is the number of factors.
The factor dimension is named, allowing for \code{out$omega1[,,"agreeableness"]}.
If \code{summary = FALSE}, then the arrays are of size [N, F, S], where S is the number of samples (\code{nsamples}).
}
\description{
Predicts omega reliability from provided data.
}
\details{
Empty for now.
}
\author{
Stephen R. Martin
}
