% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.omegad}
\alias{print.omegad}
\title{Print method for omegad objects.}
\usage{
\method{print}{omegad}(x, ...)
}
\arguments{
\item{x}{omegad object.}

\item{...}{Not used.}
}
\value{
x (invisible).
}
\description{
Print method for omegad object.
}
\details{
Prints metadata for omegad object.
}
\author{
Stephen R. Martin
}
