% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.omegad}
\alias{print.summary.omegad}
\title{Print method for summary.omegad object.}
\usage{
\method{print}{summary.omegad}(x, ...)
}
\arguments{
\item{x}{summary.omegad object.}

\item{...}{Not used.}
}
\value{
x (invisible)
}
\description{
Prints summary.omegad objects.
}
\details{
Prints summary.omegad objects.
}
\author{
Stephen R. Martin
}
