% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\alias{reliability.lavaan}
\alias{reliability.lavaan.mi}
\alias{reliability.omegad}
\alias{print.reliability.omegad}
\title{Generic function for reliability.}
\usage{
reliability(object, ...)

\method{reliability}{lavaan}(object, ...)

\method{reliability}{lavaan.mi}(object, ...)

\method{reliability}{omegad}(object, prob = 0.95, ...)

\method{print}{reliability.omegad}(x, ...)
}
\arguments{
\item{object}{omegad object.}

\item{...}{Passed onto methods.}

\item{prob}{Numeric (Default: .95). The amount of probability mass to include within the credible interval. Default values provide a 95\% credible interval.}

\item{x}{reliability.omegad object.}
}
\value{
See methods.

Matrix containing omega summaries for each factor.

x (Invisible).
}
\description{
Generic function for reliability. We include this function for compatibility with lavaan and semTools.

\code{omegad} produces observation-specific omega coefficients (See \code{\link{predict.omegad}} and \code{\link{fitted.omegad}} for extracting these).
Users may want to compute summary statistics for the omega coefficients, and to compare these to traditional estimates.
}
\details{
semTools includes a lavaan-specific reliability function. When omegad is loaded, reliability is masked.
For convenience, we define a generic \code{reliability}, with lavaan methods.
}
\section{Methods (by class)}{
\itemize{
\item \code{reliability(lavaan)}: lavaan method.

\item \code{reliability(lavaan.mi)}: lavaan.mi method.

\item \code{reliability(omegad)}: omegad method.

}}
\section{Functions}{
\itemize{
\item \code{print(reliability.omegad)}: Print method for reliability.omegad objects.

}}
\seealso{
\code{\link[semTools]{reliability}}
}
\author{
Stephen R. Martin
}
