% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.omegad}
\alias{summary.omegad}
\title{Summary method for omegad objects.}
\usage{
\method{summary}{omegad}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{omegad object.}

\item{prob}{Numeric (Default: .95). The amount of probability mass to include within the credible interval. Default values provide a 95\% credible interval.}

\item{...}{Not used.}
}
\value{
List containing summaries (Mean, SD, intervals). Dimensions provided in brackets. J = number of items, N = number of observations, F = number of factors, P = number of exogenous predictors. Items and factors are named according to the model formula:
\describe{
\item{nu_loc}{[J, 4]. Expected values (intercepts) of the indicators. (I.e., when factors are at zero.)}
\item{nu_sca}{[J, 4]. Expected residual standard deviation of the indicators. (I.e., when Error factors are at zero.)}
\item{lambda_loc_mat}{[J, 4, F]. The loadings relating the latent factor to the indicators.}
\item{lambda_sca_mat}{[J, 4, F]. The loadings relating the latent Error factor to the residual SD of the indicators.}
\item{theta_cor}{[F*2, 4, F*2] if not using a GP. [F, 4, F] if using a GP. The correlation between all factors and Error factors (when not using a GP), or between factors (when using a GP).}
\item{gp_linear}{[F, 4]. The linear coefficient between factors and their corresponding error factors.}
\item{gp_alpha}{[F, 4]. The GP "alpha" parameter between factors and their corresponding error factors. This describes the marginal SD in the Error factor due to the non-linear GP.}
\item{gp_rho}{[F, 4]. The GP "rho" parameter (i.e., length scale). This controls the "wiggliness" of the function between factors and their corresponding Error factors. The smaller it is, the more influence nearby inputs have on the function output.}
\item{exo_beta}{[P, 4, F]. The linear coefficient between exogenous predictors and the Error factors.}
\item{exo_gp_linear}{[P, 4, F]. Exogenous GP model parameter. See gp_linear.}
\item{exo_gp_alpha}{[P, 4, F]. Exogenous GP model parameter. See gp_alpha.}
\item{exo_gp_rho}{[P, 4, F]. Exogenous GP model parameter. See gp_rho.}
}
}
\description{
Summarizes omegad object.
}
\details{
Summarizes omegad object.
}
\author{
Stephen R. Martin
}
