% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.omegad}
\alias{predict.omegad}
\title{Predict method for omegad object.}
\usage{
\method{predict}{omegad}(object, newdata, summary, prob = 0.95,
  nsamples = NULL, error = TRUE, ...)
}
\arguments{
\item{object}{omegad object.}

\item{newdata}{Data.frame. Must contain the exogenous (if used) and latent factor names.}

\item{summary}{Logical (Default: TRUE). Whether to return summary or array of samples.}

\item{prob}{Numeric (Default: .95).}

\item{nsamples}{Numeric (Default: All). Number of posterior samples to use.}

\item{error}{Logical (Default: TRUE). Whether the predicted latent variables are sampled with stochastic error or not.}

\item{...}{Unused.}

\item{samples}{Numeric (Default: All). Number of posterior samples to use.}
}
\value{
Data.frame (summary=TRUE) or array (summary=FALSE).
}
\description{
Predicts omega reliability from provided data.
}
\details{
Empty for now.
}
\author{
Stephen R. Martin
}
